chore: refactor CLI
