+++ wonnx-cli/src/utils.rs
@@ -1,12 +1,14 @@
-use ndarray::ArrayBase;
+use ndarray::{Array, ArrayBase};
+use std::collections::HashMap;
 use std::path::Path;
+use wonnx::onnx::{ModelProto, TensorShapeProto, ValueInfoProto};
 use wonnx::utils::{DataTypeError, ScalarType, Shape};
 use wonnx::WonnxError;
 use wonnx_preprocessing::image::{load_bw_image, load_rgb_image};
+use wonnx_preprocessing::text::{self, BertEncodedText};
+use wonnx_preprocessing::Tensor;
 
-use wonnx::onnx::{ModelProto, TensorShapeProto, ValueInfoProto};
-
-use crate::types::NNXError;
+use crate::types::{InferOptions, InferenceInput, NNXError};
 pub trait ValueInfoProtoUtil {
     fn dimensions(&self) -> Vec<usize>;
     fn data_type(&self) -> Result<ScalarType, DataTypeError>;
@@ -124,5 +126,150 @@ pub fn load_image_input(
         }
     } else {
         Err(NNXError::InvalidInputShape)
+    }
+}
+
+impl InferenceInput {
+    pub fn new(infer_opt: &InferOptions, model: &ModelProto) -> Result<InferenceInput, NNXError> {
+        let mut inputs: HashMap<String, Tensor> = HashMap::new();
+        let mut input_shapes = HashMap::with_capacity(inputs.len());
+
+        // Do we have question and context?
+        let mut qa_encoding: Option<BertEncodedText> = None;
+        if let (Some(question), Some(context)) = (&infer_opt.question, &infer_opt.context) {
+            let tokens_input_shape = model
+                .get_input_shape(&infer_opt.qa_tokens_input)?
+                .ok_or_else(|| NNXError::InputNotFound(infer_opt.qa_tokens_input.clone()))?;
+            let mask_input_shape = model
+                .get_input_shape(&infer_opt.qa_mask_input)?
+                .ok_or_else(|| NNXError::InputNotFound(infer_opt.qa_mask_input.clone()))?;
+            let segment_input_shape = model
+                .get_input_shape(&infer_opt.qa_segment_input)?
+                .ok_or_else(|| NNXError::InputNotFound(infer_opt.qa_segment_input.clone()))?;
+
+            let segment_length = tokens_input_shape.element_count() as usize;
+
+            if segment_length != mask_input_shape.element_count() as usize {
+                return Err(NNXError::InvalidInputShape);
+            }
+            if segment_length != segment_input_shape.element_count() as usize {
+                return Err(NNXError::InvalidInputShape);
+            }
+
+            log::info!(
+                "QA: writing question '{}', context '{}' to {}/{}/{} (segment length: {})",
+                question,
+                context,
+                infer_opt.qa_tokens_input,
+                infer_opt.qa_mask_input,
+                infer_opt.qa_segment_input,
+                segment_length
+            );
+
+            let tok = text::BertTokenizer::new(Path::new(&infer_opt.vocab));
+            let encoding = tok.tokenize_question_answer(question, context)?;
+
+            let mut tokens_input = encoding.get_tokens();
+            let mut mask_input = encoding.get_mask();
+            let mut segment_input = encoding.get_segments();
+            log::debug!(
+                "tokens={:?} mask={:?} segments={:?}",
+                tokens_input,
+                mask_input,
+                segment_input
+            );
+
+            tokens_input.resize(segment_length, 0);
+            mask_input.resize(segment_length, 0);
+            segment_input.resize(segment_length, 0);
+            let tokens_input_data =
+                ndarray::Array::from_iter(tokens_input.iter().map(|x| (*x) as i64)).into_dyn();
+            let mask_input_data =
+                ndarray::Array::from_iter(mask_input.iter().map(|x| (*x) as i64)).into_dyn();
+            let segment_input_data =
+                ndarray::Array::from_iter(segment_input.iter().map(|x| (*x) as i64)).into_dyn();
+            inputs.insert(
+                infer_opt.qa_tokens_input.clone(),
+                Tensor::I64(tokens_input_data),
+            );
+            inputs.insert(
+                infer_opt.qa_mask_input.clone(),
+                Tensor::I64(mask_input_data),
+            );
+            inputs.insert(
+                infer_opt.qa_segment_input.clone(),
+                Tensor::I64(segment_input_data),
+            );
+            qa_encoding = Some(encoding);
+        }
+
+        // Process text inputs
+        if !infer_opt.text.is_empty() || !infer_opt.text_mask.is_empty() {
+            let tok = text::BertTokenizer::new(Path::new(&infer_opt.vocab));
+
+            // Tokenized text input
+            for (text_input_name, text) in &infer_opt.text {
+                let text_input_shape = model
+                    .get_input_shape(text_input_name)?
+                    .ok_or_else(|| NNXError::InputNotFound(text_input_name.clone()))?;
+                let input = tok.get_input_for(text, &text_input_shape)?;
+                inputs.insert(text_input_name.clone(), input);
+                input_shapes.insert(text_input_name.clone(), text_input_shape);
+            }
+
+            // Tokenized text input: mask
+            for (text_input_name, text) in &infer_opt.text_mask {
+                let text_input_shape = model
+                    .get_input_shape(text_input_name)?
+                    .ok_or_else(|| NNXError::InputNotFound(text_input_name.clone()))?;
+                let input = tok.get_mask_input_for(text, &text_input_shape)?;
+                inputs.insert(text_input_name.clone(), input);
+                input_shapes.insert(text_input_name.clone(), text_input_shape);
+            }
+        }
+
+        // Process raw inputs
+        for (raw_input_name, text) in &infer_opt.raw {
+            let raw_input_shape = model
+                .get_input_shape(raw_input_name)?
+                .ok_or_else(|| NNXError::InputNotFound(raw_input_name.clone()))?;
+
+            let values: Result<Vec<f32>, _> = text.split(',').map(|v| v.parse::<f32>()).collect();
+            let mut values = values.map_err(NNXError::InvalidNumber)?;
+            values.resize(raw_input_shape.element_count() as usize, 0.0);
+            inputs.insert(
+                raw_input_name.clone(),
+                Tensor::F32(Array::from_vec(values).into_dyn()),
+            );
+            input_shapes.insert(raw_input_name.clone(), raw_input_shape);
+        }
+
+        // Load input image if it was supplied
+        for (input_name, image_path) in &infer_opt.input_images {
+            let mut input_shape = model
+                .get_input_shape(input_name)?
+                .ok_or_else(|| NNXError::InputNotFound(input_name.clone()))?;
+
+            let data = load_image_input(image_path, &input_shape)?;
+
+            // Some models allow us to set the number of items we are throwing at them.
+            if input_shape.dim(0) == 0 {
+                input_shape.dims[0] = 1;
+                log::info!(
+                    "changing first dimension for input {} to {:?}",
+                    input_name,
+                    input_shape
+                );
+            }
+
+            inputs.insert(input_name.clone(), Tensor::F32(data));
+            input_shapes.insert(input_name.clone(), input_shape.clone());
+        }
+
+        Ok(InferenceInput {
+            input_shapes,
+            inputs,
+            qa_encoding,
+        })
     }
 }
